+++ nexus-app/src/main/java/com/github/nexus/key/KeyGenerator.java
@@ -1,4 +1,4 @@
-package com.github.nexus.keys;
+package com.github.nexus.key;
 
 import com.github.nexus.nacl.KeyPair;
