+++ nexus-app/src/main/java/com/github/nexus/key/KeyManager.java
@@ -1,4 +1,4 @@
-package com.github.nexus.keys;
+package com.github.nexus.key;
 
 import com.github.nexus.nacl.Key;
 import com.github.nexus.nacl.KeyPair;
