+++ nexus-app/src/main/java/com/github/nexus/key/exception/KeyNotFoundExceptionMapper.java
@@ -1,4 +1,4 @@
-package com.github.nexus.keys.exception;
+package com.github.nexus.key.exception;
 
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
