+++ nexus-app/src/main/resources/nexus-spring.xml
@@ -85,12 +85,12 @@
         <constructor-arg ref="nacl" />
     </bean>
 
-    <bean name="keyGenerator" class="com.github.nexus.keys.KeyGeneratorImpl" >
+    <bean name="keyGenerator" class="com.github.nexus.key.KeyGeneratorImpl" >
         <constructor-arg ref="nacl" />
         <constructor-arg ref="config" />
     </bean>
 
-    <bean name="keyManager" class="com.github.nexus.keys.KeyManagerImpl">
+    <bean name="keyManager" class="com.github.nexus.key.KeyManagerImpl">
         <constructor-arg ref="config" />
     </bean>
 
@@ -111,7 +111,7 @@
         
     <bean class="com.github.nexus.api.exception.DefaultExceptionMapper" />
     <bean class="com.github.nexus.api.exception.DecodingExceptionMapper"/>
-    <bean class="com.github.nexus.keys.exception.KeyNotFoundExceptionMapper"/>
+    <bean class="com.github.nexus.key.exception.KeyNotFoundExceptionMapper"/>
     <bean class="com.github.nexus.api.exception.NotFoundExceptionMapper" />
 
     <bean id="dataSource" class="org.h2.jdbcx.JdbcDataSource">
